requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2

# AI/ML
tensorflow==2.12.0
torch==2.0.1
transformers==4.29.2
nltk==3.8.1
spacy==3.5.3
sentence-transformers==2.2.2
openai==0.27.8
sounddevice==0.4.6
librosa==0.10.0.post2

# API & Web
fastapi==0.97.0
uvicorn==0.22.0
requests==2.31.0
pydantic==1.10.9
pyjwt==2.7.0
httpx==0.24.1
websockets==11.0.3
gunicorn==20.1.0
starlette==0.27.0

# Database
pymongo==4.4.0
motor==3.2.0
sqlalchemy==2.0.17
psycopg2-binary==2.9.6
alembic==1.11.1
redis==4.6.0

# Utilities
jupyter==1.0.0
pytest==7.3.1
pytest-cov==4.1.0
pytest-asyncio==0.21.0
python-dotenv==1.0.0
pyyaml==6.0
tqdm==4.65.0
loguru==0.7.0
black==23.3.0
isort==5.12.0
flake8==6.0.0
mypy==1.4.1

# Cloud & Deployment
boto3==1.26.158
google-cloud-storage==2.9.0
mlflow==2.4.1
docker==6.1.2
kubernetes==26.1.0

# Security
cryptography==41.0.1
bcrypt==4.0.1
passlib==1.7.4
